06-05-2022

+++ app/Http/Controllers/TemporaryStorageController.php
@@ -101,7 +101,7 @@ class TemporaryStorageController extends Controller
         }
         return 'ok';
     }
-    
+
 
     // remove records from db and files from folders 
     public function cleanTemporayStorage(Request $request)
@@ -118,42 +118,47 @@ class TemporaryStorageController extends Controller
     // delete removed tinyMCE images in folder and db
     public function handleTinyMceTemporaryElements(Request $request)
     {
+        // dd($request);
         $tab_data = explode(',', $request->value);
 
         // dans tiny editor on a 3 sources possibles pour le stockages des images et vidéos. 1 public/temporaryStocage, 2 public/images, 3 public/videos. Ici on supprime les images ou vidéos qui ne sont plus dans l'editor dans chacun de ces dossiers.
 
-        // delete previous images or videos from images and videos folders
-        if ($request->sender == 'collection') {
-            $description = Collection::where('id', $request->id)->first('description');
-        }
-        if ($request->sender == 'product') {
-            $description = Product::where('id', $request->id)->first('description');
+        // cette partie n'est utilisé que quand on édite
+        if ($request->id == "null") {
+            // lorsqu'on edit on récupère dabord le model édité pour effacer toutes les images et vidéos avant de save celles qui vont les remplcer
+            if ($request->sender == 'collection') {
+                $description = Collection::where('id', $request->id)->first('description');
+            }
+            if ($request->sender == 'product') {
+                $description = Product::where('id', $request->id)->first('description');
+            }
+
+            $doc = new DOMDocument();
+            @$doc->loadHTML($description);
+            $xpath = new \DOMXpath($doc);
+            $tags = $xpath->query('//img/@src | //source/@src');
+            $tab = array("\/images\/", "\/videos\/", "\\");
+            foreach ($tags as $tag) {
+                // strstr retourne une sous-chaîne allant de la première occurrence (incluse) jusqu'à la fin de la chaîne
+                $is_video = strstr($tag->value, '\/videos\/');
+                $is_image = strstr($tag->value, '\/images\/');
+                if ($is_video !== false) {
+                    $to_delete = str_replace($tab, '', $is_video);
+                    $to_delete = 'videos/' . substr($to_delete, 0, -1);
+                    if (!in_array($to_delete, $tab_data) && $tab_data[0] !== "") {
+                        if (File::exists(public_path($to_delete))) File::delete(public_path($to_delete));
+                    }
+                }
+                if ($is_image !== false) {
+                    $to_delete = str_replace($tab, '', $is_image);
+                    $to_delete = 'images/' . substr($to_delete, 0, -1);
+                    if (!in_array($to_delete, $tab_data) && $tab_data[0] !== "") {
+                        if (File::exists(public_path($to_delete))) File::delete(public_path($to_delete));
+                    }
+                }
+            }
         }
 
-        $doc = new DOMDocument();
-        @$doc->loadHTML($description);
-        $xpath = new \DOMXpath($doc);
-        $tags = $xpath->query('//img/@src | //source/@src');
-        $tab = array("\/images\/", "\/videos\/", "\\");
-        foreach ($tags as $tag) {
-            // strstr retourne une sous-chaîne allant de la première occurrence (incluse) jusqu'à la fin de la chaîne
-            $is_video = strstr($tag->value, '\/videos\/');
-            $is_image = strstr($tag->value, '\/images\/');
-            if ($is_video !== false) {
-                $to_delete = str_replace($tab, '', $is_video);
-                $to_delete = 'videos/' . substr($to_delete, 0, -1);
-                if (!in_array($to_delete, $tab_data) && $tab_data[0] !== "") {
-                    if (File::exists(public_path($to_delete))) File::delete(public_path($to_delete));
-                }
-            }
-            if ($is_image !== false) {
-                $to_delete = str_replace($tab, '', $is_image);
-                $to_delete = 'images/' . substr($to_delete, 0, -1);
-                if (!in_array($to_delete, $tab_data) && $tab_data[0] !== "") {
-                    if (File::exists(public_path($to_delete))) File::delete(public_path($to_delete));
-                }
-            }
-        }
 
         // supprime les images et vidéos dans le dossier temporaryStorage qui ne sont plus dans le tiny editor. Celles qui y sont encore sont déplacées dans leurs dossiers finaux. cette fonction est appelée dans index submit collection
         $tinyImagesVideosInDB = Temporary_storage::where('key', 'tmp_tinyMceImages')->orWhere('key', 'tmp_tinyMceVideos')->get();
